Fixes: i32/i64, broken makefile, unlink targets

+++ core/account.go
@@ -95,7 +95,7 @@ func (mw *Middleware) AccountRecover(_ *pb.RpcAccountRecoverRequest) *pb.RpcAcco
 	}
 
 	sendAccountAddEvent := func(index int, account *model.Account) {
-		m := &pb.Event{Messages: []*pb.EventMessage{{&pb.EventMessageValueOfAccountShow{AccountShow: &pb.EventAccountShow{Index: int64(index), Account: account}}}}}
+		m := &pb.Event{Messages: []*pb.EventMessage{{&pb.EventMessageValueOfAccountShow{AccountShow: &pb.EventAccountShow{Index: int32(index), Account: account}}}}}
 		if mw.SendEvent != nil {
 			mw.SendEvent(m)
 		}
